+++ app/src/main/java/com/theappbusiness/kc/utils/IntentExtras.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright (C) 2015 The App Business.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.theappbusiness.kc.utils;
+
+/**
+ * Common keys for extras we need to send from one place to another.
+ * <p>
+ * Created by voicuklein on 17/06/15.
+ */
+public class IntentExtras {
+    public static final String EXTRA_ITEM_KEY = "item";
+}
