+++ app/src/main/java/com/theappbusiness/kc/view/BaseFragment.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright (C) 2015 The App Business.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.theappbusiness.kc.view;
+
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+
+import com.theappbusiness.kc.KcApp;
+
+public abstract class BaseFragment extends Fragment {
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        // Perform injection so that when this call returns, all dependencies will be available for use.
+        ((KcApp) getActivity().getApplication()).inject(this);
+    }
+}
