+++ app/src/main/res/menu/menu_main_test.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.nabilhachicha.kc.items.itemlist.MainTestActivity">
+    tools:context="com.theappbusiness.kc.items.itemlist.MainTestActivity">
     <item android:id="@+id/action_settings" android:title="@string/action_settings"
         android:orderInCategory="100" android:showAsAction="never" />
 </menu>
